Slight expansion on Factory design pattern amongst Actors

+++ src/main/java/actors/Car.java
@@ -2,7 +2,7 @@ package actors;
 
 public class Car extends Obstacle {
 	
-	public Car (String filename, int w, int h, int xpos, int ypos, int s) {
+	public Car (String filename, int w, int h, int xpos, int ypos, double s) {
 		super(filename, w, h, xpos, ypos, s);
 		
 //		Image car1Left = new Image(ActorImages.IMG_CAR_LEFT_1, w, h, true, true);
