+++ src/main/java/actors/Truck.java
@@ -2,7 +2,7 @@ package actors;
 
 public class Truck extends Obstacle {
 	
-	public Truck (String filename, int w, int h, int xpos, int ypos, int s) {
+	public Truck (String filename, int w, int h, int xpos, int ypos, double s) {
 		super(filename, w, h, xpos, ypos, s);
 		
 	}
